fixed capitalization on the landing page
I fixed the capitalization for the landing page and added translations for the account activation. Are html-characters possible in the way I used them? (I deliberately chose two different arrows, to check how it looks)

In `Constants.ts` i commented in a suggestion for a fixed number format, that doesn't require internationalization and is currently the most adopted in the scientific community, because it "always" happens that there are mixups with "." and "," which leads to all sorts of pains.