Labs can change availability status of user
Added two new endpoints, so that labs can update the availability status of a volunteer. The volunteer is informed by e-mail about this, with the possibility to discard or accept the status change. After 3 days without reaction to the mail, the availability status of the volunteer is set to "not available".

* Backend completed
* Refactored code a little
   * Moved Database models to separate file
   * Migrated database cronjob to separate file